Changed achievers, improved MILP
The low-level actions that have no mappings must share the same
resources as their higher-level action, hence they are going to be
constrained by them. This may change in the future to integrate colored
links representing the constraints.

MILP is now its own package so that it can be called from the main
script. Good luck with packages in Python