Fixed achievers, removed achievers check for low-level actions
The achiever function should now check correctly if an action is an
achiever of another action. Also I included some tests in
./prolog_planner/utility/test/conditions_tests.pl that can be used
to check if the functions are working properly.

As for the low-level actions, looking for achievers between them it's
tricky as they tend to have negative pre-conditions such as
!gripping(a,_), which always match on effects such as
del(gripping(a,block1)), making it virtually imposible to differentiate
between predicates that are actually needed and not. One possibility
would be to not just check whether HPre and del(HPre) match, but to
check both the name of the predicate and the arguments position-wise.

For the moment, this should be sufficient though.